The script is able to export ideal beams/columns
The model is successfully connected in column nodes as lines. Beams are idealized using the top middle longest line and columns in the middle center along the axis.

The model is exported to two .scr files one containing lines that represent 3d shapes and the other with lines representing the idealized elements.

Gotta work on exporting a JSON file readable by FEM.js