Started to work on JSON exporting
Created a script to work on the JSON exporting. By now, have nodes in a dict with tag and coords, beams in a dict with initial node tag and end one, same for columns.

The algorithm is not the fastest since look for all the nodes until the tag is found.